+++ R/Defaults.R
@@ -34,17 +34,21 @@
 
 # List of the default parsing functions. The names of the list are the
 # names of the respective XML documents set in Constants.R.
+# For SOS 1.0.0, the matching is done via operation name.
+# For SOS 2.0.0, the matching is done via root name of the response document.
 .createDefaultParsers <- function() {
   .defP <- list(
     parseSosCapabilities,
-    parseSensorML,
-    parseOM,
-    parseGetObservationResponse,
-    parseGetFeatureOfInterestResponse,
-    parseOM,
     parseOwsExceptionReport,
     #
+    parseSensorML,
+    parseDescribeSensorResponse,
+    #
+    parseOM,
+    parseOM,
+    #
     parseGetDataAvailabilityResponse,
+    parseGetFeatureOfInterestResponse,
     parseGetObservationByIdResponse,
     parseGetObservationResponse,
     #
@@ -75,16 +79,18 @@
 
   names(.defP) <- list(
     sosGetCapabilitiesName,
-    sosDescribeSensorName,
-    sosGetObservationName,
-    sosGetObservationResponseName,
-    sosGetFeatureOfInterestResponseName,
-    sosGetObservationByIdName,
     owsExceptionReportName,
     #
+    sosDescribeSensorName,
+    swesDescribeSensorResponseName,
+    #
+    sosGetObservationName,
+    sosGetObservationByIdName,
+    #
     gdaGetDataAvailabilityResponseName,
-    sos200GetObservationByIdResponseName,
-    sos200GetObservationResponseName,
+    sosGetFeatureOfInterestResponseName,
+    sosGetObservationByIdResponseName,
+    sosGetObservationResponseName,
     #
     omMeasurementName,
     omMemberName,